Move index set to shared
